# Improve SpanHelpers.IndexOfAny(byte,byte)