The better approached used in https://github.com/dotnet/runtime/pull/40729

```
| Faster                                                   | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 16)   |      1.45 |             5.02 |             3.45 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 24)   |      1.52 |             5.93 |             3.91 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 32)   |      1.44 |             6.09 |             4.23 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 48)   |      1.69 |             5.82 |             3.45 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 64)   |      1.40 |             5.82 |             4.17 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 128)  |      1.29 |             6.36 |             4.94 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 512)  |      1.43 |            10.75 |             7.54 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 256)  |      1.34 |             7.58 |             5.67 |         |
| System.Memory.Span<Byte>.IndexOfAnyTwoValues(Size: 1024) |      1.11 |            12.99 |            11.69 |         |
```

Will effect [header parsing](https://github.com/dotnet/aspnetcore/blob/ca1505ba3b0b0717ff07f8df88443cfe214acd5c/src/Servers/Kestrel/Core/src/Internal/Http/HttpParser.cs#L205) in Kestrel
```csharp
length = span.IndexOfAny(ByteCR, ByteLF)
```

/cc @adamsitnik @danmosemsft