@danmosemsft 2020-08-14
@benaadams @sebastienros is there a way to quantify how much it would help Kestrel? Would help prioritize review.