@benaadams 2020-08-14
Top one of the SpanHelpers for PlainText Platform on Citrine on Linux

![image](https://user-images.githubusercontent.com/1142958/90210722-acd3ce00-dde6-11ea-956f-7834a80cee43.png)

Not sure how it changes it though :)