@danmosemsft 2020-08-14
@sebastienros do you expect this to materially improve Plaintext, given data above?