@sebastienros 2020-08-14
We can hope it doesn't make it slower! Ideally any perf related work should be measured before merged, even more that close to rc. We'll have to look at the charts once the build is available. We can also do a custom with the next build containing this change. I'll follow the CI.